feat: Task 2 - Create Entity and Attribute tables

+++ task_2.sql
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS Order_Details (
     order_details_id INT AUTO_INCREMENT PRIMARY KEY,
     order_id INT,
     book_id INT,
-    quantity INT NOT NULL,
+    quantity DOUBLE NOT NULL,
     unit_price DECIMAL(10, 2),
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
